refactor: 정규식 상수화 및 불필요한 유효성검사 파일 삭제
