+++ src/validation/cardOwner.ts
@@ -1,5 +1,7 @@
+import { REGULAR_EXPRESSION } from "../CONSTANT";
+
 export const validateCardOwner = (owner: string) => {
-  const regExp = /^[A-Z\s]+$/;
+  const regExp = REGULAR_EXPRESSION.CAPITAL_LETTER_OR_SPACE;
 
   return regExp.test(owner);
 };
