+++ src/validation/number.ts
@@ -1,5 +1,7 @@
+import { REGULAR_EXPRESSION } from "../CONSTANT";
+
 export const validateNumber = (value: string) => {
-  const regExp = /^\d{1}$/;
+  const regExp = REGULAR_EXPRESSION.ONE_NUMBER;
 
   return regExp.test(value);
 };
