--- src/validation/password.ts
@@ -1,4 +1,0 @@
-export const validatePassword = (value: string) => {
-  const regExp = /^\d{0,1}$/;
-  return regExp.test(value);
-};
