--- src/validation/securityNumber.ts
@@ -1,4 +1,0 @@
-export const validateSecurityNumber = (value: string) => {
-  const regExp = /^\d{0,3}$/;
-  return regExp.test(value);
-};
